+++ src/gt4py/next/embedded/common.py
@@ -15,6 +15,8 @@ from __future__ import annotations
 
 from typing import Any, Optional, Sequence, cast
 
+import numpy as np
+
 from gt4py.next import common
 from gt4py.next.embedded import exceptions as embedded_exceptions
 
@@ -144,3 +146,15 @@ def broadcast_domain(
                 (dim, common.UnitRange(common.Infinity.negative(), common.Infinity.positive()))
             )
     return named_ranges
+
+
+def _compute_domain_slice(
+    field: common.Field, new_dimensions: tuple[common.Dimension, ...]
+) -> Sequence[slice | None]:
+    domain_slice: list[slice | None] = []
+    for dim in new_dimensions:
+        if find_index_of_dim(dim, field.domain) is not None:
+            domain_slice.append(slice(None))
+        else:
+            domain_slice.append(np.newaxis)
+    return domain_slice
